Update README.md
1. This file needs to be copied to the solver directory inside electromagnetics sub-folder.
2. Once copied, compile the solver and it is ready to use.

Note: This is a transient solver for MHD turbulent flow with only minor changes to mhdFoam. For certain problems, use of the solver can become very time consuming. People are more than welcome to make changes to the solver and make a steady state solver for it. This would really save a lot of time!